- renamed some JS files to TS - moved Vitest config to Vite config - deleted Vitest config file - updated Vite aliases - added typing to main.ts
