--- vite.config.js
@@ -1,18 +1,0 @@
-import path from 'path'
-import { defineConfig } from 'vite'
-import vue from '@vitejs/plugin-vue2'
-import EnvironmentPlugin from 'vite-plugin-environment'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [
-    vue(),
-    EnvironmentPlugin({ BUILD: 'web' }) // fix for Vuelidate
-  ],
-  resolve: {
-    alias: {
-      '@': path.resolve(__dirname, './src')
-    },
-    extensions: ['.mjs', '.js', '.ts', '.jsx', '.tsx', '.json', '.vue']
-  }
-})
